requirements.txt
# Core dependencies
aiohttp>=3.8.0
requests>=2.28.0
pyyaml>=6.0
tqdm>=4.65.0

# Utilities
click>=8.1.0          # CLI interface
colorama>=0.4.6       # Colored output
python-dotenv>=1.0.0  # Environment variables

# Development dependencies
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.3.0         # Code formatting
flake8>=6.0.0         # Linting
mypy>=1.3.0           # Type checking

# Optional dependencies
# huggingface-hub>=0.16.0  # For direct HF API access
# civitai>=0.1.0           # For Civitai API (if available)